removed sample.py
